OGEMA 2.1.1 public release

+++ src/core/ref-impl/resource-manager/src/main/java/org/ogema/resourcemanager/impl/model/schedule/DefaultSchedule.java
@@ -16,6 +16,7 @@
 package org.ogema.resourcemanager.impl.model.schedule;
 
 import java.util.Collection;
+import java.util.Iterator;
 import java.util.List;
 
 import org.ogema.core.application.ApplicationManager;
@@ -28,6 +29,7 @@ import org.ogema.core.timeseries.InterpolationMode;
 import org.ogema.persistence.impl.faketree.ScheduleTreeElement;
 import org.ogema.resourcemanager.impl.ApplicationResourceManager;
 import org.ogema.resourcemanager.impl.ResourceBase;
+import org.ogema.resourcemanager.impl.ResourceDBManager;
 import org.ogema.resourcemanager.virtual.VirtualTreeElement;
 
 /**
@@ -41,13 +43,16 @@ public class DefaultSchedule extends ResourceBase implements org.ogema.core.mode
 		org.ogema.core.model.schedule.ForecastSchedule, AbsoluteSchedule, RelativeSchedule {
 
 	private final ApplicationManager m_appMan;
+	// FIXME this needs to change when a reference is set or replaced; need a ScheduleTreeElementRegistry!
 	private final ScheduleTreeElement m_scheduleElement;
+	private final ResourceDBManager m_dbMan;
 
 	public DefaultSchedule(VirtualTreeElement el, ScheduleTreeElement scheduleElement, String path,
-			ApplicationResourceManager resMan, ApplicationManager appMan) {
+			ApplicationResourceManager resMan, ApplicationManager appMan, ResourceDBManager dbMan) {
 		super(el, path, resMan);
 		m_appMan = appMan;
 		m_scheduleElement = scheduleElement;
+		m_dbMan = dbMan;
 	}
 
 	final ScheduleTreeElement getSchedule() {
@@ -58,21 +63,33 @@ public class DefaultSchedule extends ResourceBase implements org.ogema.core.mode
 	@SuppressWarnings("unchecked")
 	public <T extends Resource> T create() {
 		getResourceDB().lockStructureWrite();
+		m_dbMan.startTransaction();
 		try {
 			DefaultSchedule s = super.create();
 			m_scheduleElement.create();
 			return (T) s;
 		} finally {
+			m_dbMan.finishTransaction();
 			getResourceDB().unlockStructureWrite();
 		}
 	}
 
     @Override
     protected void deleteTreeElement() {
-        getSchedule().deleteValues();
-        getSchedule().setLastModified(-1);
-        m_scheduleElement.getScheduleElement().delete();
-        super.deleteTreeElement();
+    	m_dbMan.startTransaction();
+		try {
+			if (!isReference(false)) {
+		        getSchedule().deleteValues();
+		        getSchedule().setLastModified(-1);
+		        m_scheduleElement.getScheduleElement().delete();
+			} else {
+				// TODO this is not working!
+			}
+	        super.deleteTreeElement();
+		} finally {
+			m_dbMan.finishTransaction();
+		}
+		
     }
 
 	@Override
@@ -81,7 +98,12 @@ public class DefaultSchedule extends ResourceBase implements org.ogema.core.mode
 			return false;
 		}
 		checkWritePermission();
-		getSchedule().addValue(timestamp, value);
+		m_dbMan.startTransaction();
+		try {
+			getSchedule().addValue(timestamp, value);
+		} finally {
+			m_dbMan.finishTransaction();
+		}
 		getSchedule().setLastUpdateTime(m_appMan.getFrameworkTime());
 		handleResourceUpdate(true);
 		return true;
@@ -93,7 +115,12 @@ public class DefaultSchedule extends ResourceBase implements org.ogema.core.mode
 			return false;
 		}
 		checkWritePermission();
-		getSchedule().addValues(values);
+		m_dbMan.startTransaction();
+		try {
+			getSchedule().addValues(values);
+		} finally {
+			m_dbMan.finishTransaction();
+		}
 		getSchedule().setLastUpdateTime(m_appMan.getFrameworkTime());
 		handleResourceUpdate(true);
 		return true;
@@ -111,7 +138,12 @@ public class DefaultSchedule extends ResourceBase implements org.ogema.core.mode
 			return false;
 		}
 		checkWritePermission();
-		getSchedule().replaceValuesFixedStep(startTime, values, stepSize);
+		m_dbMan.startTransaction();
+		try {
+			getSchedule().replaceValuesFixedStep(startTime, values, stepSize);
+		} finally {
+			m_dbMan.finishTransaction();
+		}
 		getSchedule().setLastUpdateTime(m_appMan.getFrameworkTime());
 		handleResourceUpdate(true);
 		return true;
@@ -123,7 +155,12 @@ public class DefaultSchedule extends ResourceBase implements org.ogema.core.mode
 			return false;
 		}
 		checkWritePermission();
-		getSchedule().addValue(timestamp, value, timeOfCalculation);
+		m_dbMan.startTransaction();
+		try {
+			getSchedule().addValue(timestamp, value, timeOfCalculation);
+		} finally {
+			m_dbMan.finishTransaction();
+		}
 		getSchedule().setLastUpdateTime(m_appMan.getFrameworkTime());
 		handleResourceUpdate(true);
 		return true;
@@ -135,7 +172,12 @@ public class DefaultSchedule extends ResourceBase implements org.ogema.core.mode
 			return false;
 		}
 		checkWritePermission();
-		getSchedule().addValues(values, timeOfCalculation);
+		m_dbMan.startTransaction();
+		try {
+			getSchedule().addValues(values, timeOfCalculation);
+		} finally {
+			m_dbMan.finishTransaction();
+		}
 		getSchedule().setLastUpdateTime(m_appMan.getFrameworkTime());
 		handleResourceUpdate(true);
 		return true;
@@ -153,7 +195,12 @@ public class DefaultSchedule extends ResourceBase implements org.ogema.core.mode
 			return false;
 		}
 		checkWritePermission();
-		getSchedule().replaceValuesFixedStep(startTime, values, stepSize, timeOfCalculation);
+		m_dbMan.startTransaction();
+		try {
+			getSchedule().replaceValuesFixedStep(startTime, values, stepSize, timeOfCalculation);
+		} finally {
+			m_dbMan.finishTransaction();
+		}
 		getSchedule().setLastUpdateTime(m_appMan.getFrameworkTime());
 		handleResourceUpdate(true);
 		return true;
@@ -170,7 +217,12 @@ public class DefaultSchedule extends ResourceBase implements org.ogema.core.mode
 			return false;
 		}
 		checkWritePermission();
-		getSchedule().deleteValues();
+		m_dbMan.startTransaction();
+		try {
+			getSchedule().deleteValues();
+		} finally {
+			m_dbMan.finishTransaction();
+		}
 		getSchedule().setLastUpdateTime(m_appMan.getFrameworkTime());
 		handleResourceUpdate(true);
 		return true;
@@ -182,7 +234,12 @@ public class DefaultSchedule extends ResourceBase implements org.ogema.core.mode
 			return false;
 		}
 		checkWritePermission();
-		getSchedule().deleteValues(endTime);
+		m_dbMan.startTransaction();
+		try {
+			getSchedule().deleteValues(endTime);
+		} finally {
+			m_dbMan.finishTransaction();
+		}
 		getSchedule().setLastUpdateTime(m_appMan.getFrameworkTime());
 		handleResourceUpdate(true);
 		return true;
@@ -194,7 +251,12 @@ public class DefaultSchedule extends ResourceBase implements org.ogema.core.mode
 			return false;
 		}
 		checkWritePermission();
-		getSchedule().deleteValues(startTime, endTime);
+		m_dbMan.startTransaction();
+		try {
+			getSchedule().deleteValues(startTime, endTime);
+		} finally {
+			m_dbMan.finishTransaction();
+		}
 		getSchedule().setLastUpdateTime(m_appMan.getFrameworkTime());
 		handleResourceUpdate(true);
 		return true;
@@ -206,7 +268,12 @@ public class DefaultSchedule extends ResourceBase implements org.ogema.core.mode
 			return false;
 		}
 		checkWritePermission();
-		getSchedule().replaceValues(startTime, endTime, values);
+		m_dbMan.startTransaction();
+		try {
+			getSchedule().replaceValues(startTime, endTime, values);
+		} finally {
+			m_dbMan.finishTransaction();
+		}
 		getSchedule().setLastUpdateTime(m_appMan.getFrameworkTime());
 		handleResourceUpdate(true);
 		return true;
@@ -257,6 +324,41 @@ public class DefaultSchedule extends ResourceBase implements org.ogema.core.mode
 	@Override
 	public long getLastUpdateTime() {
 		return getSchedule().getLastModified();
+	}
+	
+	@Override
+	public SampledValue getPreviousValue(long time) {
+		return getSchedule().getPreviousValue(time);
+	}
+
+	@Override
+	public boolean isEmpty() {
+		return getSchedule().isEmpty();
+	}
+
+	@Override
+	public boolean isEmpty(long startTime, long endTime) {
+		return getSchedule().isEmpty(startTime, endTime);
+	}
+
+	@Override
+	public int size() {
+		return getSchedule().size();
+	}
+
+	@Override
+	public int size(long startTime, long endTime) {
+		return getSchedule().size(startTime, endTime);
+	}
+
+	@Override
+	public Iterator<SampledValue> iterator() {
+		return getSchedule().iterator();
+	}
+
+	@Override
+	public Iterator<SampledValue> iterator(long startTime, long endTime) {
+		return getSchedule().iterator(startTime, endTime);
 	}
 
 }
